Display permissions form
This commit display the permissions form on roles pages.
The form is operating as espected, but the manner how permissions will
be sent to the backend will differ from the previous version.

You can manage each permission, by checking the checkboxes on each line:
  - show: let users having this role to display that kind of elements,
  - add: let users having this role to create new elements of this kind,
  - edit: let users having this role to edit that kind of elements,
  - delete: let users having this role to delete that kind of elements,
  - all: a magical checkbox, that let you check or uncheck all
    checkboxes of the line.

This commit also improve displaying of the checkboxes so that it take
less space on mobile.